+++ Configuration/TCA/Overrides/tt_content.php
@@ -97,3 +97,51 @@ $GLOBALS['TCA']['tt_content']['types']['themebootstrap4_text'] = [
         ],
     ],
 ];
+//
+// Simple media content element
+$GLOBALS['TCA']['tt_content']['columns']['CType']['config']['items'][] = [
+    'Media',
+    'themebootstrap4_media',
+    'new-content-el-media',
+];
+$GLOBALS['TCA']['tt_content']['ctrl']['typeicon_classes']['themebootstrap4_media'] = 'new-content-el-media';
+$GLOBALS['TCA']['tt_content']['types']['themebootstrap4_media'] = [
+    'showitem' => '
+            CType,
+            header,
+            bodytext,
+        --div--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:tabs.media,
+            assets,
+        --div--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:tabs.appearance,
+            --palette--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:palette.frames;frames,
+        --div--;LLL:EXT:core/Resources/Private/Language/Form/locallang_tabs.xlf:access,
+            --palette--;;hidden,
+            --palette--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:palette.access;access,
+        --div--;LLL:EXT:gridelements/Resources/Private/Language/locallang_db.xlf:gridElements,
+            tx_gridelements_container,
+            tx_gridelements_columns,
+            colPos',
+    'columnsOverrides' => [
+        'layout' => [
+            'config' => [
+                'type' => 'select',
+                'renderType' => 'selectSingle',
+                'minitems' => 1,
+                'maxitems' => 1,
+                'size' => 1,
+                'items' => [
+                    ['Default', '0'],
+                    ['Media-Element links', 'MediaLeft'],
+                    ['Media-Element rechts', 'MediaRight'],
+                    ['Media-Element unten', 'MediaBottom'],
+                ],
+                'eval' => 'trim',
+            ],
+        ],
+        'bodytext' => [
+            'config' => [
+                'enableRichtext' => 1,
+            ],
+        ],
+    ],
+];